python-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv
langchain
langchain-community
faiss-cpu
sentence-transformers
langchain-groq
pypdf
PyPDF2==3.0.1